move map to dataframe.py
