+++ src/snowflake/snowpark/dataframe.py
@@ -10,9 +10,11 @@ import sys
 from collections import Counter
 from functools import cached_property
 from logging import getLogger
+from types import ModuleType
 from typing import (
     TYPE_CHECKING,
     Any,
+    Callable,
     Dict,
     Iterator,
     List,
@@ -23,7 +25,7 @@ from typing import (
 )
 
 import snowflake.snowpark
-from snowflake.connector.options import installed_pandas
+from snowflake.connector.options import installed_pandas, pandas
 from snowflake.snowpark._internal.analyzer.binary_plan_node import (
     AsOf,
     Cross,
@@ -104,6 +106,7 @@ from snowflake.snowpark._internal.type_utils import (
     LiteralType,
     snow_type_to_dtype_str,
 )
+from snowflake.snowpark._internal.udf_utils import add_package_to_existing_packages
 from snowflake.snowpark._internal.utils import (
     SKIP_LEVELS_THREE,
     SKIP_LEVELS_TWO,
@@ -157,7 +160,13 @@ from snowflake.snowpark.table_function import (
     _get_cols_after_explode_join,
     _get_cols_after_join_table,
 )
-from snowflake.snowpark.types import StringType, StructType, _NumericType
+from snowflake.snowpark.types import (
+    PandasDataFrameType,
+    StringType,
+    StructField,
+    StructType,
+    _NumericType,
+)
 
 # Python 3.8 needs to use typing.Iterable because collections.abc.Iterable is not subscriptable
 # Python 3.9 can use both
@@ -4335,3 +4344,217 @@ Query List:
     # joinTableFunction = join_table_function
     # naturalJoin = natural_join
     # withColumns = with_columns
+
+
+def map(
+    dataframe: DataFrame,
+    func: Callable,
+    output_types: List[StructType],
+    *,
+    output_column_names: Optional[List[str]] = None,
+    imports: Optional[List[Union[str, Tuple[str, str]]]] = None,
+    packages: Optional[List[Union[str, ModuleType]]] = None,
+    immutable: bool = False,
+    partition_by: Optional[Union[ColumnOrName, List[ColumnOrName]]] = None,
+    vectorized: bool = False,
+    max_batch_size: Optional[int] = None,
+):
+    """Returns a new DataFrame with the result of applying `func` to each of the
+    rows of the specified DataFrame.
+
+    This function registers a temporary `UDTF
+    <https://docs.snowflake.com/en/developer-guide/udf/python/udf-python-tabular-functions>`_ and
+    returns a new DataFrame with the result of applying the `func` function to each row of the
+    given DataFrame.
+
+    Args:
+        dataframe: The DataFrame instance.
+        func: A function to be applied to every row of the DataFrame.
+        output_types: A list of types for values generated by the ``func``
+        output_column_names: A list of names to be assigned to the resulting columns.
+        imports: A list of imports that are required to run the function. This argument is passed
+            on when registering the UDTF.
+        packages: A list of packages that are required to run the function. This argument is passed
+            on when registering the UDTF.
+        immutable: A flag to specify if the result of the func is deterministic for the same input.
+        partition_by: Specify the partitioning column(s) for the UDTF.
+        vectorized: A flag to determine if the UDTF process should be vectorized. See
+            `vectorized UDTFs <https://docs.snowflake.com/en/developer-guide/udf/python/udf-python-tabular-vectorized#udtfs-with-a-vectorized-process-method>`_.
+        max_batch_size: The maximum number of rows per input pandas DataFrame when using vectorized option.
+
+    Example 1::
+
+        >>> from snowflake.snowpark.types import IntegerType
+        >>> from snowflake.snowpark.dataframe import map
+        >>> import pandas as pd
+        >>> df = session.create_dataframe([[10, "a", 22], [20, "b", 22]], schema=["col1", "col2", "col3"])
+        >>> new_df = map(df, lambda row: row[0] * row[0], output_types=[IntegerType()])
+        >>> new_df.order_by("c_1").show()
+        ---------
+        |"C_1"  |
+        ---------
+        |100    |
+        |400    |
+        ---------
+        <BLANKLINE>
+
+    Example 2::
+
+        >>> new_df = map(df, lambda row: (row[1], row[0] * 3), output_types=[StringType(), IntegerType()])
+        >>> new_df.order_by("c_1").show()
+        -----------------
+        |"C_1"  |"C_2"  |
+        -----------------
+        |a      |30     |
+        |b      |60     |
+        -----------------
+        <BLANKLINE>
+
+    Example 3::
+
+        >>> new_df = map(
+        ...     df,
+        ...     lambda row: (row[1], row[0] * 3),
+        ...     output_types=[StringType(), IntegerType()],
+        ...     output_column_names=['col1', 'col2']
+        ... )
+        >>> new_df.order_by("col1").show()
+        -------------------
+        |"COL1"  |"COL2"  |
+        -------------------
+        |a       |30      |
+        |b       |60      |
+        -------------------
+        <BLANKLINE>
+
+    Example 4::
+
+        >>> new_df = map(df, lambda pdf: pdf['COL1']*3, output_types=[IntegerType()], vectorized=True)
+        >>> new_df.order_by("c_1").show()
+        ---------
+        |"C_1"  |
+        ---------
+        |30     |
+        |60     |
+        ---------
+        <BLANKLINE>
+
+    Example 5::
+
+        >>> new_df = map(
+        ...     df,
+        ...     lambda pdf: (pdf['COL1']*3, pdf['COL2']+"b"),
+        ...     output_types=[IntegerType(), StringType()],
+        ...     output_column_names=['A', 'B'],
+        ...     vectorized=True
+        ... )
+        >>> new_df.order_by("A").show()
+        -------------
+        |"A"  |"B"  |
+        -------------
+        |30   |ab   |
+        |60   |bb   |
+        -------------
+        <BLANKLINE>
+
+    Example 6::
+
+        >>> new_df = map(
+        ...     df,
+        ...     lambda pdf: ((pdf.shape[0],) * len(pdf), (pdf.shape[1],) * len(pdf)),
+        ...     output_types=[IntegerType(), IntegerType()],
+        ...     output_column_names=['rows', 'cols'],
+        ...     partition_by="col3",
+        ...     vectorized=True,
+        ... )
+        >>> new_df.show()
+        -------------------
+        |"ROWS"  |"COLS"  |
+        -------------------
+        |2       |3       |
+        |2       |3       |
+        -------------------
+        <BLANKLINE>
+
+    Note:
+        - The result of the `func` function must be either a scalar value or a tuple containing the same number of elements
+        as specified in the `output_types` argument.
+        - When using the `vectorized` option, the `func` function must accept a pandas DataFrame as input and return either
+        a pandas DataFrame, or a tuple of pandas Series/arrays.
+
+    """
+    if len(output_types) == 0:
+        raise ValueError("output_types cannot be empty.")
+
+    if output_column_names is None:
+        output_column_names = [f"c_{i+1}" for i in range(len(output_types))]
+    elif len(output_column_names) != len(output_types):
+        raise ValueError(
+            "'output_column_names' and 'output_types' must be of the same size."
+        )
+
+    df_columns = dataframe.columns
+    packages = packages or list(dataframe._session.get_packages().values())
+    packages = add_package_to_existing_packages(packages, "snowflake-snowpark-python")
+
+    input_types = [field.datatype for field in dataframe.schema.fields]
+    udtf_output_cols = [f"c{i}" for i in range(len(output_types))]
+    output_schema = StructType(
+        [StructField(col, type_) for col, type_ in zip(udtf_output_cols, output_types)]
+    )
+
+    if vectorized:
+        # If the map is vectorized, we need to add pandas to packages if not
+        # already added. Also update the input_types and output_schema to
+        # be PandasDataFrameType.
+        packages = add_package_to_existing_packages(packages, pandas)
+        input_types = [PandasDataFrameType(input_types)]
+        output_schema = PandasDataFrameType(output_types, udtf_output_cols)
+
+    output_columns = [
+        col(f"${i + len(df_columns) + 1}").alias(
+            col_name
+        )  # this is done to avoid collision with original table columns
+        for i, col_name in enumerate(output_column_names)
+    ]
+
+    def wrap_result(result):
+        if vectorized:
+            if isinstance(result, pandas.DataFrame) or isinstance(result, tuple):
+                return result
+            return (result,)
+
+        if isinstance(result, Row):
+            return tuple(result)
+        elif isinstance(result, tuple):
+            return result
+        else:
+            return (result,)
+
+    if vectorized:
+
+        class _MapFunc:
+            def process(self, pdf):
+                return wrap_result(func(pdf))
+
+    else:
+
+        class _MapFunc:
+            def process(self, *argv):
+                input_args_to_row = Row(*df_columns)
+                yield wrap_result(func(input_args_to_row(*argv)))
+
+    map_udtf = dataframe._session.udtf.register(
+        _MapFunc,
+        output_schema=output_schema,
+        input_types=input_types,
+        input_names=df_columns,
+        imports=imports,
+        packages=packages,
+        immutable=immutable,
+        max_batch_size=max_batch_size,
+    )
+
+    return dataframe.join_table_function(
+        map_udtf(*df_columns).over(partition_by=partition_by)
+    ).select(*output_columns)
