+++ tests/integ/test_dataframe.py
@@ -15,6 +15,7 @@ from itertools import product
 from typing import Tuple
 from unittest import mock
 
+from snowflake.snowpark.dataframe import map
 from snowflake.snowpark.session import Session
 
 try:
@@ -4206,3 +4207,160 @@ def test_dataframe_to_local_iterator_with_to_pandas_isolation(
     # local testing always give 1 chunk
     if not local_testing_mode:
         assert batch_count > 1
+
+
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.udf
+@pytest.mark.parametrize("overlapping_columns", [True, False])
+@pytest.mark.parametrize(
+    "func,output_types,output_col_names,expected",
+    [
+        (lambda row: row[1] + 1, [IntegerType()], ["A"], [(i + 1,) for i in range(5)]),
+        (
+            lambda row: (row.B * 2, row.C),
+            [IntegerType(), StringType()],
+            ["B", "C"],
+            [
+                (
+                    i * 2,
+                    f"w{i}",
+                )
+                for i in range(5)
+            ],
+        ),
+        (
+            lambda row: Row(row.B * row.B, f"-{row.C}"),
+            [IntegerType(), StringType()],
+            ["B", "C"],
+            [(i * i, f"-w{i}") for i in range(5)],
+        ),
+    ],
+)
+def test_map_basic(
+    session, func, output_types, output_col_names, expected, overlapping_columns
+):
+    df = session.create_dataframe(
+        [
+            (
+                True,
+                i,
+                f"w{i}",
+            )
+            for i in range(5)
+        ],
+        schema=["A", "B", "C"],
+    )
+
+    if overlapping_columns:
+        row = Row(*output_col_names)
+        expected = [row(*e) for e in expected]
+        Utils.check_answer(
+            map(df, func, output_types, output_column_names=output_col_names), expected
+        )
+    else:
+        row = Row(*[f"c_{i+1}" for i in range(len(output_types))])
+        expected = [row(*e) for e in expected]
+        Utils.check_answer(map(df, func, output_types), expected)
+
+
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.skipif(not is_pandas_available, reason="pandas is required for this test")
+@pytest.mark.udf
+@pytest.mark.parametrize(
+    "func,output_types,expected",
+    [
+        (lambda df: df["B"] + 1, [IntegerType()], [(i + 1,) for i in range(5)]),
+        (
+            lambda df: pd.concat([df["B"] * 2, df["C"]], axis=1),
+            [IntegerType(), StringType()],
+            [
+                (
+                    i * 2,
+                    f"w{i}",
+                )
+                for i in range(5)
+            ],
+        ),
+    ],
+)
+def test_map_vectorized(session, func, output_types, expected):
+    df = session.create_dataframe(
+        [
+            (
+                True,
+                i,
+                f"w{i}",
+            )
+            for i in range(5)
+        ],
+        schema=["A", "B", "C"],
+    )
+
+    Utils.check_answer(
+        map(df, func, output_types, vectorized=True, packages=["pandas"]), expected
+    )
+
+
+@pytest.mark.skipif(
+    "config.getoption('local_testing_mode', default=False)",
+    reason="Table function is not supported in Local Testing",
+)
+@pytest.mark.udf
+def test_map_chained(session):
+    df = session.create_dataframe(
+        [[True, i, f"w{i}"] for i in range(5)], schema=["A", "B", "C"]
+    )
+
+    new_df = map(
+        map(
+            df,
+            lambda x: (x.B * x.B, f"_{x.C}_"),
+            output_types=[IntegerType(), StringType()],
+        ),
+        lambda x: len(x[1]) + x[0],
+        output_types=[IntegerType()],
+    )
+    expected = [(len(f"_w{i}_") + i * i,) for i in range(5)]
+
+    Utils.check_answer(new_df, expected)
+
+    # chained calls with repeated column names
+    new_df = map(
+        map(
+            df,
+            lambda x: Row(x.B * x.B, f"_{x.C}_"),
+            output_types=[IntegerType(), StringType()],
+            output_column_names=["A", "B"],
+        ),
+        lambda x: Row(len(x.B) + x.A),
+        output_types=[IntegerType()],
+        output_column_names=["A"],
+        packages=["snowflake-snowpark-python"],
+    )
+    Utils.check_answer(new_df, expected)
+
+
+def test_map_negative(session):
+    df1 = session.create_dataframe(
+        [[True, i, f"w{i}"] for i in range(5)], schema=["A", "B", "C"]
+    )
+
+    with pytest.raises(ValueError, match="output_types cannot be empty."):
+        map(df1, lambda row: [row.B, row.C], output_types=[])
+
+    with pytest.raises(
+        ValueError,
+        match="'output_column_names' and 'output_types' must be of the same size.",
+    ):
+        map(
+            df1,
+            lambda row: [row.B, row.C],
+            output_types=[IntegerType(), StringType()],
+            output_column_names=["a", "b", "c"],
+        )
